+++ src/Logging/src/Abstractions/DynamicConsoleLogger.cs
