+++ src/Logging/src/DynamicSerilogBase/SerilogLoggingBuilderExtensions.cs
@@ -6,26 +6,27 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Configuration;
-using Microsoft.Extensions.Logging.Console;
-using Microsoft.Extensions.Options;
 using Serilog;
-using Serilog.Core;
-using Serilog.Extensions.Logging;
 using System;
 using System.Linq;
-using static Steeltoe.Extensions.Logging.DynamicLoggingBuilder;
 
 namespace Steeltoe.Extensions.Logging.DynamicSerilog
 {
     public static class SerilogLoggingBuilderExtensions
     {
-        public static ILoggingBuilder AddDynamicSerilog(this ILoggingBuilder builder, IConfiguration configuration, LoggerConfiguration serilogConfiguration = null, bool preserveDefaultConsole = false)
+        public static ILoggingBuilder AddDynamicSerilog(this ILoggingBuilder builder, LoggerConfiguration serilogConfiguration, bool preserveDefaultConsole = false)
         {
             if (builder == null)
             {
                 throw new ArgumentNullException(nameof(builder));
             }
+
+            if (builder.Services.Any(sd => sd.ServiceType == typeof(IDynamicLoggerProvider)))
+            {
+                throw new InvalidOperationException("An IDynamicLoggerProvider has already been configured! Call 'AddDynamicSerilog' earlier in program.cs (before adding Actuators) or remove duplicate IDynamicLoggerProvider entries.");
+            }
+
+            builder.AddFilter<SerilogDynamicProvider>(null, LogLevel.Trace);
 
             // only run if an IDynamicLoggerProvider hasn't already been added
             if (!builder.Services.Any(sd => sd.ServiceType == typeof(IDynamicLoggerProvider)))
@@ -35,16 +36,16 @@ namespace Steeltoe.Extensions.Logging.DynamicSerilog
                     builder.ClearProviders();
                 }
 
-                builder.AddFilter<SerilogDynamicProvider>(null, LogLevel.Trace); // TODO : What does this do?
-                builder.Services.AddOptions();
-
                 if (serilogConfiguration != null)
                 {
-                    builder.Services.Configure<SerilogOptions>(options => options.SetSerilogOptions(serilogConfiguration));
+                    builder.Services.AddSingleton(serilogConfiguration);
+                    builder.Services.AddOptions<SerilogOptions>()
+                        .Configure<LoggerConfiguration>((options, serilogConfiguration) => options.SetSerilogOptions(serilogConfiguration));
                 }
                 else
                 {
-                    builder.Services.Configure<SerilogOptions>(options => options.SetSerilogOptions(configuration));
+                    builder.Services.AddOptions<SerilogOptions>()
+                        .Configure<IConfiguration>((options, iconfiguration) => options.SetSerilogOptions(iconfiguration));
                 }
 
                 builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<ILoggerProvider, SerilogDynamicProvider>());
@@ -63,8 +64,7 @@ namespace Steeltoe.Extensions.Logging.DynamicSerilog
         /// <returns>The configured <see cref="ILoggingBuilder"/></returns>
         public static ILoggingBuilder AddDynamicSerilog(this ILoggingBuilder builder, bool preserveDefaultConsole = false)
         {
-            var configuration = builder.Services.BuildServiceProvider().GetService<IConfiguration>();
-            return builder.AddDynamicSerilog(configuration, null, preserveDefaultConsole);
+            return builder.AddDynamicSerilog(null, preserveDefaultConsole);
         }
 
         /// <summary>
@@ -76,11 +76,10 @@ namespace Steeltoe.Extensions.Logging.DynamicSerilog
         /// <param name="preserveStaticLogger">Not supported!</param>
         /// <param name="preserveDefaultConsole">When true, do not remove Microsoft's ConsoleLoggerProvider</param>
         /// <returns>The configured <see cref="ILoggingBuilder"/></returns>
-        /// TODO: Obsolete
-        public static ILoggingBuilder AddDynamicSerilog(this ILoggingBuilder builder, LoggerConfiguration loggerConfiguration, bool preserveStaticLogger = false, bool preserveDefaultConsole = false)
+        [Obsolete("Please use a different overload of AddDynamicSerilog ")]
+        public static ILoggingBuilder AddDynamicSerilog(this ILoggingBuilder builder, LoggerConfiguration loggerConfiguration, bool preserveStaticLogger, bool preserveDefaultConsole = false)
         {
-            var configuration = builder.Services.BuildServiceProvider().GetService<IConfiguration>();
-            return builder.AddDynamicSerilog(configuration, loggerConfiguration, preserveDefaultConsole);
+            return builder.AddDynamicSerilog(loggerConfiguration, preserveDefaultConsole);
         }
 
         /// <summary>
@@ -90,7 +89,7 @@ namespace Steeltoe.Extensions.Logging.DynamicSerilog
         /// <param name="builder">The <see cref="ILoggingBuilder"/> for configuring the LoggerFactory  </param>
         /// <returns>The configured <see cref="ILoggingBuilder"/></returns>
         [Obsolete("Please use AddDynamicSerilog instead")]
-        public static ILoggingBuilder AddSerilogDynamicConsole(this ILoggingBuilder builder) => builder.AddDynamicSerilog();
+        public static ILoggingBuilder AddSerilogDynamicConsole(this ILoggingBuilder builder) => builder.AddDynamicSerilog(false);
 
         /// <summary>
         /// Add Steeltoe logger wrapped in a <see cref="IDynamicLoggerProvider"/> that supports
