+++ src/Logging/src/DynamicSerilogBase/Steeltoe/Extensions/Logging/DynamicLoggingBuilder.cs
@@ -1,0 +1,10 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information.
+
+namespace Steeltoe.Extensions.Logging
+{
+    internal class DynamicLoggingBuilder
+    {
+    }
+}
