+++ src/Logging/src/DynamicSerilogCore/SerilogWebHostBuilderExtensions.cs
@@ -15,32 +15,45 @@ namespace Steeltoe.Extensions.Logging.DynamicSerilog
         /// </summary>
         /// <param name="hostBuilder">The <see cref="IWebHostBuilder"/> to configure</param>
         /// <param name="configureLogger">The delegate for configuring the <see cref="DynamicLoggerConfiguration" /> that will be used to construct a <see cref="Serilog.Core.Logger" /></param>
-        /// <param name="preserveStaticLogger">Not Supported!</param>
         /// <param name="preserveDefaultConsole">When true, do not remove Microsoft's ConsoleLoggerProvider</param>
         /// <returns>The <see cref="IWebHostBuilder"/></returns>
         public static IWebHostBuilder AddDynamicSerilog(
             this IWebHostBuilder hostBuilder,
             Action<WebHostBuilderContext, LoggerConfiguration> configureLogger = null,
-            bool preserveStaticLogger = false,
             bool preserveDefaultConsole = false)
         {
             return hostBuilder
-                .ConfigureLogging((hostContext, logBuilder) =>
-                {
-                    LoggerConfiguration loggerConfiguration = null;
-                    if (configureLogger is object)
-                    {
-                        loggerConfiguration = new LoggerConfiguration().ReadFrom.Configuration(hostContext.Configuration);
-                        configureLogger(hostContext, loggerConfiguration);
+               .ConfigureLogging((hostContext, logBuilder) =>
+               {
+                   LoggerConfiguration loggerConfiguration = null;
+                   if (configureLogger is object)
+                   {
+                       loggerConfiguration = new LoggerConfiguration().ReadFrom.Configuration(hostContext.Configuration);
+                       configureLogger(hostContext, loggerConfiguration);
 
-                        logBuilder.AddDynamicSerilog(loggerConfiguration, preserveStaticLogger, preserveDefaultConsole);
-                    }
-                    else
-                    {
-                        logBuilder.AddDynamicSerilog(hostContext.Configuration, null, false);
-                    }
-                });
+                       logBuilder.AddDynamicSerilog(loggerConfiguration, preserveDefaultConsole);
+                   }
+                   else
+                   {
+                       logBuilder.AddDynamicSerilog(null, false);
+                   }
+               });
         }
+
+        /// <summary>
+        /// Configure Serilog as the <see cref="IDynamicLoggerProvider"/> to enable dynamically controlling log levels via management endpoints
+        /// </summary>
+        /// <param name="hostBuilder">The <see cref="IWebHostBuilder"/> to configure</param>
+        /// <param name="configureLogger">The delegate for configuring the <see cref="DynamicLoggerConfiguration" /> that will be used to construct a <see cref="Serilog.Core.Logger" /></param>
+        /// <param name="preserveStaticLogger">Not Supported!</param>
+        /// <param name="preserveDefaultConsole">When true, do not remove Microsoft's ConsoleLoggerProvider</param>
+        /// <returns>The <see cref="IWebHostBuilder"/></returns>
+        [Obsolete("Use a different overload of AddDynamicSerilog")]
+        public static IWebHostBuilder AddDynamicSerilog(
+            this IWebHostBuilder hostBuilder,
+            Action<WebHostBuilderContext, LoggerConfiguration> configureLogger,
+            bool preserveStaticLogger,
+            bool preserveDefaultConsole = false) => hostBuilder.AddDynamicSerilog(configureLogger, preserveDefaultConsole);
 
         /// <summary>
         /// Sets Steeltoe <see cref="IDynamicLoggerProvider"/> Serilog implementation as a LoggerProvider which supports
